refactoring presque terminé, reste seulement des petits bug a corriger
