+++ sae_5_2/controllers/GridFrameController.py
@@ -1,0 +1,2 @@
+class GridFrameController :
+    pass
