+++ sae_5_2/controllers/LeftNavbarController.py
@@ -1,0 +1,2 @@
+class LeftNavbarController:
+    pass
