+++ sae_5_2/controllers/TopNavbarController.py
@@ -1,0 +1,2 @@
+class TopNavbarController:
+    pass
