+++ sae_5_2/views/GridCanvas.py
@@ -1,6 +1,6 @@
-import customtkinter as ctk
+import tkinter as tk
 
-class GridCanvas(ctk.CTkCanvas):
+class GridCanvas(tk.Canvas):
     def __init__(self, parent):
         super().__init__(parent, bg="grey")
